+++ NEWS.md
@@ -12,12 +12,14 @@
 - Export a new `prune_graph()` function for igraph objects.
 - Delete long-deprecated functions `prune()` and `status()`.
 - Deprecate and rename functions:
+    - `analyses()` => `plan_analyses()`
     - `backend()` => `future_backend()`
     - `check()` => `check_plan()`
     - `evaluate()` => `evaluate_plan()`
     - `expand()` => `expand_plan()`
     - `gather()` => `gather_plan()`
     - `session()` => `drake_session()`
+    - `summaries()` => `plan_summaries()`
 - Disallow `output` and `code` as names in the workflow plan data frame. Use `target` and `command` instead. This naming switch has been formally deprecated for several months prior.
 
 # Changes in release 4.4.0